Condense code into files based on day
* Move HistorianList class into day1.py
* Move constants into the file they are used
