--- advent_of_code/constants.py
@@ -1,2 +1,0 @@
-INPUT_DAY1_PATH = "advent_of_code/input_day1.txt"
-INPUT_DAY2_PATH = "advent_of_code/input_day2.txt"
