+++ advent_of_code/day2.py
@@ -2,16 +2,18 @@
 Advent of code 2024 day 2.
 """
 
-from advent_of_code import constants, util
+from advent_of_code import util
 from typing import List
 from copy import deepcopy
+
+INPUT_DAY2_PATH = "advent_of_code/input_day2.txt"
 
 
 def main():
     util.print_output_string(2, 1)
-    print(count_safe_reports(constants.INPUT_DAY2_PATH))
+    print(count_safe_reports(INPUT_DAY2_PATH))
     util.print_output_string(2, 2)
-    print(count_safe_reports(constants.INPUT_DAY2_PATH, True))
+    print(count_safe_reports(INPUT_DAY2_PATH, True))
 
 
 def count_safe_reports(filepath: str, allow_bad_level=False) -> int:
@@ -57,28 +59,26 @@ def is_report_safe(report: List[str], allow_bad_level=False) -> bool:
                     return False
                 # Check if report is safe with any of these items removed
                 return (
-                    is_report_safe_without_item_at_index(report, i - 1)
-                    or is_report_safe_without_item_at_index(report, i)
-                    or is_report_safe_without_item_at_index(report, i + 1)
+                    is_report_safe_without_item(report, i - 1)
+                    or is_report_safe_without_item(report, i)
+                    or is_report_safe_without_item(report, i + 1)
                 )
 
         # Compare previous and current level difference
         if not difference_is_between_one_and_three(prev_lvl, current_lvl):
             if not allow_bad_level:
                 return False
-            return is_report_safe_without_item_at_index(report, i - 1) or is_report_safe_without_item_at_index(
-                report, i
-            )
+            return is_report_safe_without_item(report, i - 1) or is_report_safe_without_item(report, i)
 
     # Finally compare last two levels not compared in loop
     if difference_is_between_one_and_three(int(report[-1]), int(report[-2])):
         return True
     elif allow_bad_level:
-        return is_report_safe_without_item_at_index(report, -1) or is_report_safe_without_item_at_index(report, -2)
+        return is_report_safe_without_item(report, -1) or is_report_safe_without_item(report, -2)
     return False
 
 
-def is_report_safe_without_item_at_index(report: List[str], index: int) -> bool:
+def is_report_safe_without_item(report: List[str], index: int) -> bool:
     """Returns True if the given report would be safe if the item at the given
     index was removed, False otherwise.
     """
